restrição de campo e criação de exception

+++ src/main/java/br/com/alelo/consumer/consumerpat/vo/BuyVo.java
@@ -15,7 +15,7 @@ public class BuyVo {
 
     private Long cardNumber;
     private BigDecimal value;
-    private Long establishmentId;
+    private Integer establishmentId;
     private Long establishmentType;
     private String productDescription;
